🚚 파일 이름 변경
sub-issues:
- [NAILCASE-393](https://nailcase.atlassian.net/browse/NAILCASE-393)
