+++ src/component/custom/manager/base/my-shop/home/banner/index.tsx
@@ -8,6 +8,7 @@ import type { TResGetShopInfo } from "@/util/api_v2/get-shop-Info"
 
 import NTContent from "../../../../../../common/nt-content/index"
 
+import { BannerTitle } from "./banner-title"
 import { getEssestialImageProps } from "./util"
 
 export default function ManagerBanner() {
@@ -37,21 +38,6 @@ export default function ManagerBanner() {
 				{`${selectedIdx + 1}/${profileImageLength.toString()}`}
 			</NTContent>
 			<BannerTitle InfoData={InfoData} />
-		</div>
-	)
-}
-type BannerTitlePT = {
-	InfoData: TResGetShopInfo
-}
-function BannerTitle({ InfoData }: BannerTitlePT) {
-	return (
-		<div className="absolute left-[90px] top-[95px]">
-			<p className="text-Callout text-[14px] font-Light text-White">
-				네일아트 전문 | {InfoData.address}
-			</p>
-			<h1 className="text-Title01 text-[28px] font-Bold text-White">
-				{InfoData.shopName}
-			</h1>
 		</div>
 	)
 }
